requirements.txt
folium==0.15.1
haversine==2.8.1
numpy==1.25.2
pandas==2.0.3
Pillow==10.3.0
plotly==5.16.1
streamlit==1.34.0
streamlit_folium==0.18.0